sepalate files
