+++ srcs/t3_render.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   t3_render.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: corvvs <corvvs@student.42.fr>              +#+  +:+       +#+        */
+/*   By: tkomatsu <tkomatsu@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/02/15 21:44:13 by corvvs            #+#    #+#             */
-/*   Updated: 2022/02/19 01:54:03 by corvvs           ###   ########.fr       */
+/*   Created: 2022/02/20 16:02:43 by tkomatsu          #+#    #+#             */
+/*   Updated: 2022/02/20 16:02:45 by tkomatsu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
