+++ srcs/t3_reset_params.c
@@ -1,0 +1,19 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   t3_reset_params.c                                  :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tkomatsu <tkomatsu@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/02/20 16:03:01 by tkomatsu          #+#    #+#             */
+/*   Updated: 2022/02/20 16:03:02 by tkomatsu         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "term3d.h"
+
+void	t3_reset_params(t_system *system)
+{
+	dprintf(STDERR_FILENO, "reset parameters\n");
+	t3_init_render_params(system);
+}
