+++ srcs/t3_ui.c
@@ -75,100 +75,6 @@ int	t3_get_key(void)
 	return (0);
 }
 
-void	t3_update_omega(t_system *system, int key)
-{
-	if (key == 'e' || key == 'q')
-	{
-		system->optics.omega += ((key == 'e') * 2 - 1) * 0.000001;
-		dprintf(
-			STDERR_FILENO,
-			"omega -> %f[rot/s]\n",
-			system->optics.omega * 1000000 / M_PI / 2
-			);
-	}
-}
-
-void	t3_update_offset(t_system *system, int key)
-{
-	if (key == 'd' || key == 'a')
-	{
-		system->optics.offset_x += ((key == 'd') * 2 - 1) * 0.1;
-		dprintf(STDERR_FILENO, "offset.x -> %f\n", system->optics.offset_x);
-	}
-	else if (key == 's' || key == 'w')
-	{
-		system->optics.offset_y += ((key == 's') * 2 - 1) * 0.1;
-		dprintf(STDERR_FILENO, "offset.y -> %f\n", system->optics.offset_x);
-	}
-}
-
-void	t3_update_flame(t_system *system, int key)
-{
-	if (key == 'x' || key == 'z')
-	{
-		if ((key == 'x' && system->optics.fps < 240)
-			|| (key == 'z' && 2 < system->optics.fps))
-			system->optics.fps += ((key == 'x') * 2 - 1);
-		system->optics.uspf = 1000000 / system->optics.fps;
-		dprintf(STDERR_FILENO, "fps -> %f\n", system->optics.fps);
-	}
-}
-
-void	t3_update_scale(t_system *system, int key)
-{
-	if (key == '1')
-	{
-		system->optics.scale_factor /= 1.05;
-		dprintf(STDERR_FILENO, "scale -> %f\n", system->optics.scale_factor);
-	}
-	else if (key == '2')
-	{
-		system->optics.scale_factor *= 1.05;
-		dprintf(STDERR_FILENO, "scale -> %f\n", system->optics.scale_factor);
-	}
-}
-
-void	t3_udpate_axis(t_system *system, int key)
-{
-	t_vector3d	*axis;
-
-	axis = &system->optics.rot_axis;
-	if (key == '3')
-	{
-		t3_rot_vector_around_axis(system->optics.rot_axis,
-			(t_vector3d){1, 0, 0}, M_PI / 1024);
-		dprintf(STDERR_FILENO, "rot axis -> (%f, %f, %f)\n",
-			*axis[0], *axis[1], *axis[2]);
-	}
-	else if (key == '4')
-	{
-		t3_rot_vector_around_axis(system->optics.rot_axis,
-			(t_vector3d){0, 1, 0}, M_PI / 1024);
-		dprintf(STDERR_FILENO, "rot axis -> (%f, %f, %f)\n",
-			*axis[0], *axis[1], *axis[2]);
-	}
-	else if (key == '5')
-	{
-		t3_rot_vector_around_axis(system->optics.rot_axis,
-			(t_vector3d){0, 0, 1}, M_PI / 1024);
-		dprintf(STDERR_FILENO, "rot axis -> (%f, %f, %f)\n",
-			*axis[0], *axis[1], *axis[2]);
-	}
-}
-
-void	t3_update_animate(t_system *system)
-{
-	system->optics.animate = !system->optics.animate;
-	dprintf(STDERR_FILENO, "animate -> %s\n",
-		(char *[2]){"NO", "YES"}[!!system->optics.animate]);
-}
-
-void	t3_reset_params(t_system *system)
-{
-	dprintf(STDERR_FILENO, "reset parameters\n");
-	t3_init_render_params(system);
-}
-
 void	t3_update_by_key(t_system *system)
 {
 	const int	key = t3_get_key();
